openVideoInGallery moved to videoUtils
